+++ VTK/Utilities/vtkexodus2/exgssc.c
@@ -52,7 +52,7 @@
 *                                       side or face
 * revision history - 
 *
-*  $Id: exgssc.c,v 1.4 2006-11-29 18:09:13 dcthomp Exp $
+*  $Id: exgssc.c,v 1.5 2006-11-29 20:33:02 dcthomp Exp $
 *****************************************************************************/
 
 #include <ctype.h>
@@ -300,7 +300,7 @@ int ex_get_side_set_node_count(int exoid,
     elem_blk_parms[i].num_nodes_per_elem = num_nodes_per_elem;
     elem_blk_parms[i].num_attr = num_attr;
 
-    for (m=0; m < strlen(elem_type); m++) {
+    for (m=0; m < (int)strlen(elem_type); m++) {
       elem_blk_parms[i].elem_type[m] = toupper((int)elem_type[m]);
     }
     elem_blk_parms[i].elem_type[m] = '\0';
